// ClientApp/src/app/models/associated-me-no-stock-comment.ts
export class AssociatedMeNoStockComment {
    ME_PART_NUMBER: string;
    KEYWORD: string;
    STATION: string;
    ON_HAND: number;
    ALLOCATION: number;
    ASSOCIATED_ON_HAND: number;
    ASSOCIATED_ALLOCATION: number;
    CAMPUS_ON_HAND: number;
    DAYSINQUEUE: number;
    COMM: string;
    CHKKEY: string;
}